requirements.txt
Flask==1.1.1
gunicorn==20.0.4
slackclient==2.5.0
firebase-admin==4.0.0
python-dotenv==0.12.0